Changed to long's solution wide.

+++ pf_re/pf_re/line.c
@@ -2,24 +2,24 @@
 
 #include "line.h"
 
-Line *newLineH (int i)
+Line *newLineH (long i)
 {
 	Line *line	= (Line*)safeMalloc (sizeof (Line));
-	int x_orig = i%(m-1);
-	int y = (i/(m-1));
-	int x_dest = x_orig+1;
+	long x_orig = i%(m-1);
+	long y = (i/(m-1));
+	long x_dest = x_orig+1;
 	line->origin = getNode(x_orig,y);
 	line->destination = getNode(x_dest, y);
 	line->mine = 0;	
 
 	return line;
 }
-Line *newLineV (int i)
+Line *newLineV (long i)
 {
 	Line *line	= (Line*)safeMalloc (sizeof (Line));
-	int x = i%(m);
-	int y_orig = (i/(m));
-	int y_dest = y_orig+1;
+	long x = i%(m);
+	long y_orig = (i/(m));
+	long y_dest = y_orig+1;
 	line->origin = getNode(x,y_orig);
 	line->destination = getNode(x, y_dest);
 	line->mine = 0;	
@@ -29,8 +29,8 @@ Line *newLineV (int i)
 Line *getLine(Node *origin, Node *destination){
 	return getLineFilter(origin, destination, 1);
 }
-Line *getLineFilter(Node *origin, Node *destination, int filter){
-	int i;
+Line *getLineFilter(Node *origin, Node *destination, long filter){
+	long i;
 	for(i=0; i<numLines; i++){
 		if((lines[i]->origin == origin && lines[i]->destination == destination) || (lines[i]->origin == destination && lines[i]->destination == origin) && (lines[i]->mine == 0||!filter))
 			return lines[i];
@@ -39,7 +39,7 @@ Line *getLineFilter(Node *origin, Node *destination, int filter){
 }
 
 
-int placeMine(Node *origin, Node *destination){
+long placeMine(Node *origin, Node *destination){
 	Line *line = getLine(origin,destination);
 	if(line){
 		line->mine = 1;
@@ -48,7 +48,7 @@ int placeMine(Node *origin, Node *destination){
 	return 0;
 }
 
-int removeMine(Node *origin, Node *destination){
+long removeMine(Node *origin, Node *destination){
 	Line *line = getLine(origin,destination);
 	if(line){
 		line->mine = 0;
